requirements.txt
certifi==2023.7.22
matplotlib==3.7.2
numpy==1.21.1
filterpy
opencv-python
pandas==1.2
pyparsing
PyYAML
scikit-image
seaborn
six
sympy
tqdm==4.64.0
tzdata
scipy==1.8
typing-extensions
ultralytics==8.0.114
urllib3
python-dotenv==1.0.1
